+++ src/main/java/com/lockertracker/repository/LockerRepository.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Repository;
 import com.lockertracker.model.LockerDBModel;
 
 @Repository
-public interface LockerRepository extends JpaRepository<LockerDBModel, Short> {
+public interface LockerRepository extends JpaRepository<LockerDBModel, Long> {
 
-	public LockerDBModel findById(final short id);
+	public LockerDBModel findById(final long id);
 
 }
